Updating light logic trackign
Instead of just using an if-then statement, updates a queue to track what lights will show up in order.

switchLight() removed in favor of updateQueue(), and light-switching functionality moved to progress().

The object now tracks the last meaningful state (any state other than all red lights) and fills its queue based on what that last light was.

It still needs some hardcoding -- the state & lastState attribute determines the next light behavior and the state attribute determines the behavior of 1 light from now.  We should consider basing this on user input in the future.