Gaussian car generation
Previous car generation had a somewhat-normal distribution, but in a way that could be easily described.

Replaced this with a Gaussian distribution (with tails cut off at +/- 1) which initializes 100 values at the start of the simulation and then references those cyclically throughout the rest of the simulation.

This models the behavior of traffic in a way that is (hopefully) easier to troubleshoot. Maximum wait time for a car is now 2x the frequency variable.

This will eventually begin to lap itself, but the time at which that happens is so far away that it's no longer a risk.